file :: generateToken
