+++ src/service/auth.ts
@@ -1,5 +1,5 @@
 import { signUp } from './auth/signup';
-import { refresh } from './auth/token';
+import { refresh } from './auth/refresh';
 import { login } from './auth/login';
 import { logout } from './auth/logout';
