+++ src/utils/jwt.ts
@@ -12,3 +12,15 @@ export const signJWT = (payload: object, expiresIn: ms.StringValue | number): st
   };
   return jwt.sign(payload, privateKey, signOptions);
 };
+
+export const generateToken = async (id: string, isAccess: boolean) => {
+  const token = signJWT(
+    {
+      id,
+      type: isAccess ? 'access' : 'refresh',
+      iat: Math.floor(Date.now() / 1000)
+    },
+    isAccess ? '1h' : '7d'
+  );
+  return token;
+};
